M19 matchmaking: fix editor compile errors
- MatchmakingApi.MatchItem was missing WhiteSteamID/BlackSteamID/OpponentName,
  which the server's matched-row poll returns and Matchmaking.SeatMatchAsHost reads.
- SetupPanel BuildHash had a 9-arg HashCode.Combine (max 8); nest the tail.
- NameFor: cast SteamId->ulong explicitly (matches HostHandleDisconnect), avoiding
  an ambiguous == overload.

+++ client/Code/World/LobbyNetworkManager.cs
@@ -200,7 +200,7 @@ public sealed class LobbyNetworkManager : Component, Component.INetworkListener,
 	string NameFor( ulong steamId )
 	{
 		foreach ( var c in _conns )
-			if ( c.SteamId == steamId ) return c.DisplayName;
+			if ( (ulong)c.SteamId == steamId ) return c.DisplayName;
 		return "Player";
 	}
